Changed SolarSystem service WARNING: file operations were removed.
